Moved solve to separate file
